CLI command to generate components
By default, this will generate a placeholder Twig template and
Context file, each containing all the properties passed to the
command via the --properties argument.

By passing the --controller flag, it will also generate a
placeholder Controller class in the core plugin.

To skip the template, use the --no-template flag.

To skip the context, use the --no-context flag.

Any slashes in the component name will be used to build
the path to the component, and will be used in the namespace
of the Context and Controller classes.